fix(backend): continue refactor
