+++ backend/endpoints/get/get_donations.go
@@ -1,4 +1,4 @@
-package endpoints
+package get
 
 import (
 	"net/http"
